requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
cffi==1.15.1
h11==0.14.0
idna==3.4
outcome==1.2.0
pip==23.0.1
pycparser==2.21
PySocks==1.7.1
selenium==4.8.2
setuptools==67.5.1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
trio-websocket==0.9.2
trio==0.22.0
urllib3==1.26.14
wheel==0.38.4
wsproto==1.2.0
pyodbc~=4.0.35
openpyxl~=3.1.1
requests~=2.28.2
screeninfo~=0.8.1
pandas~=1.5.3
python-dotenv~=1.0.0
pymongo~=4.3.3
pika~=1.3.1